requirements.txt
sounddevice
wavio
pydub
google-generativeai
playwright
python-dotenv